+++ src/rules/index.ts
@@ -8,6 +8,7 @@ import { unknownFieldRule } from "./unknown-field.js";
 import { unstableReferenceRule } from "./unstable-reference.js";
 import { dangerousCommandRule } from "./dangerous-command.js";
 import { httpWithoutAuthRule } from "./http-without-auth.js";
+import { duplicateEnvKeyRule } from "./duplicate-env-key.js";
 
 export const BUILTIN_RULES: Rule[] = [
   structureRules,
@@ -19,4 +20,5 @@ export const BUILTIN_RULES: Rule[] = [
   unstableReferenceRule,
   dangerousCommandRule,
   httpWithoutAuthRule,
+  duplicateEnvKeyRule,
 ];